requirements.txt
Flask==2.3.2
pdfplumber==0.10.2
python-docx==0.8.11
speechrecognition==3.10.0
pydub==0.25.1
numpy==1.26.0
scikit-learn==1.3.0
requests==2.31.0
pyannote.audio
torch
torchvision
torchaudio
python-dotenv
docx
openai
sarvamai
Flask-SQLAlchemy
Flask-Migrate
mysqlclient